Add HTTP 1.x client configurations
