+++ src/test/resources/http-client.yml
@@ -1,0 +1,4 @@
+connectionTimeout: 600ms
+idleTimeout: 3m
+connectionFactory:
+  type : 'http'
